+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 0.6.1
 
+- __bug__: `className` prop passed to `React` and `Svelte` components is now applied to the outer div
 - __bug__: fixes `transition` still active even if it was set to `false`
 - __bug__: fixes type validation of `transition` `props` in `React`, `VueJS 2.x` and `VueJS 3.x`
 - __chore__: removes irrelevant attribute `className` in angular